requirement.txt
Django-2.1.5
djangorestframework-3.9.0 
drf-nested-routers-0.91